castParam accepts a raw openapi shape
